test: Added tests to cover the added lines.
