+++ tests/sapientml/test_generatedcode.py
@@ -20,6 +20,7 @@ from pathlib import Path
 from shutil import copyfile
 
 import nbformat
+import numpy as np
 import pandas as pd
 import pytest
 from sapientml.executor import run
@@ -975,3 +976,119 @@ def test_misc_timeout_works_hyperparameter_tuning_timeout(
     pipeline_results = execute_pipeline(dataset, task, config, temp_dir, initial_timeout=initial_timeout)
 
     assert pipeline_results[0][1].returncode == 0
+
+
+@pytest.mark.parametrize("invalid_target_value", [np.nan, np.inf, -np.inf])
+@pytest.mark.parametrize("invalid_count", ["half", "one"])
+def test_nan_or_inf_in_target_columns_for_train_data(
+    invalid_target_value,
+    invalid_count,
+    setup_request_parameters,
+    make_tempdir,
+    execute_pipeline,
+    test_df_train,
+    test_df_valid,
+    test_df_test,
+):
+    task, config, dataset = setup_request_parameters()
+
+    # test pattern setting
+    n_models = 1
+    config.n_models = n_models
+
+    task.task_type = "regression"
+    task.adaptation_metric = "RMSLE"
+    task.target_columns = ["target_number_large_scale"]
+
+    dataset.training_dataframe = test_df_train
+    dataset.validation_dataframe = test_df_valid
+    dataset.test_dataframe = test_df_test
+
+    dataset.training_data_path = (fxdir / "datasets" / "testdata_train.csv").as_posix()
+    dataset.validation_data_path = (fxdir / "datasets" / "testdata_valid.csv").as_posix()
+    dataset.test_data_path = (fxdir / "datasets" / "testdata_test.csv").as_posix()
+
+    temp_dir = make_tempdir
+    if invalid_count == "half":
+        dataset.training_dataframe.loc[0::2, task.target_columns] = invalid_target_value
+        with pytest.raises(ValueError) as e:
+            execute_pipeline(dataset, task, config, temp_dir, initial_timeout=60)
+        assert "dataframe has more than 10 rows including NaN or Inf." in str(e.value)
+    else:
+        dataset.training_dataframe.loc[dataset.training_dataframe.index[0], task.target_columns] = invalid_target_value
+        pipeline_results = execute_pipeline(dataset, task, config, temp_dir, initial_timeout=60)
+        assert not (pipeline_results is None)
+
+
+@pytest.mark.parametrize("invalid_target_value", [np.nan, np.inf, -np.inf])
+def test_nan_or_inf_in_target_columns_for_validation_data(
+    invalid_target_value,
+    setup_request_parameters,
+    make_tempdir,
+    execute_pipeline,
+    test_df_train,
+    test_df_valid,
+    test_df_test,
+):
+    task, config, dataset = setup_request_parameters()
+
+    # test pattern setting
+    n_models = 1
+    config.n_models = n_models
+
+    task.task_type = "regression"
+    task.adaptation_metric = "RMSLE"
+    task.target_columns = ["target_number_large_scale"]
+
+    dataset.training_dataframe = test_df_train
+    dataset.validation_dataframe = test_df_valid
+    dataset.test_dataframe = test_df_test
+
+    dataset.training_data_path = (fxdir / "datasets" / "testdata_train.csv").as_posix()
+    dataset.validation_data_path = (fxdir / "datasets" / "testdata_valid.csv").as_posix()
+    dataset.test_data_path = (fxdir / "datasets" / "testdata_test.csv").as_posix()
+
+    dataset.validation_dataframe.loc[dataset.validation_dataframe.index[0], task.target_columns] = invalid_target_value
+
+    temp_dir = make_tempdir
+    with pytest.raises(ValueError) as e:
+        execute_pipeline(dataset, task, config, temp_dir, initial_timeout=60)
+
+    assert "dataframe has NaN or Inf" in str(e.value)
+
+
+@pytest.mark.parametrize("invalid_target_value", [np.nan, np.inf, -np.inf])
+def test_nan_or_inf_in_target_columns_for_test_data(
+    invalid_target_value,
+    setup_request_parameters,
+    make_tempdir,
+    execute_pipeline,
+    test_df_train,
+    test_df_valid,
+    test_df_test,
+):
+    task, config, dataset = setup_request_parameters()
+
+    # test pattern setting
+    n_models = 1
+    config.n_models = n_models
+
+    task.task_type = "regression"
+    task.adaptation_metric = "RMSLE"
+    task.target_columns = ["target_number_large_scale"]
+
+    dataset.training_dataframe = test_df_train
+    dataset.validation_dataframe = test_df_valid
+    dataset.test_dataframe = test_df_test
+
+    dataset.training_data_path = (fxdir / "datasets" / "testdata_train.csv").as_posix()
+    dataset.validation_data_path = (fxdir / "datasets" / "testdata_valid.csv").as_posix()
+    dataset.test_data_path = (fxdir / "datasets" / "testdata_test.csv").as_posix()
+
+    dataset.test_dataframe.loc[dataset.test_dataframe.index[0], task.target_columns] = invalid_target_value
+
+    temp_dir = make_tempdir
+    with pytest.raises(ValueError) as e:
+        execute_pipeline(dataset, task, config, temp_dir, initial_timeout=60)
+
+    assert "dataframe has NaN or Inf" in str(e.value)
